LibWeb: Split PercentageStyleValue out of StyleValue.{h,cpp}
